updated. made more flexable

+++ spatial_join_areas.py
@@ -14,7 +14,7 @@ def create_database_connection(db_params):
     return connection
 
 
-def main(connection=None, connection_parameters=None, dataset='', dataset_area=True, areas_to_join='', fields_to_join=[]):
+def main(connection=None, connection_parameters=None, dataset='', join_multiple_areas=True, areas_to_join='', join_data_is_areas=True, fields_to_join=[], dataset_id_field='gid', join_areas_id_field='gid', dataset_geom_field='geom', join_areas_geom_field='geom'):
     """
     Run the spatial join over two tables existing in a postgres database.
 
@@ -32,32 +32,41 @@ def main(connection=None, connection_parameters=None, dataset='', dataset_area=T
 
     temp_table = '___temp'
 
-    # run join
-    if dataset_area:
-        cursor.execute(sql.SQL('SELECT b.gid, ARRAY_AGG(st_area(st_intersection(b.geom, t.geom)) / st_area(b.geom)) as coverage, ARRAY_AGG(t.geo_code) as lads, ARRAY_AGG(t.geo_code_gor) as gors INTO {} FROM ftables.{} t, {} b WHERE st_intersects(b.geom, t.geom) ORDER BY g.gid;').format(sql.Identifier(temp_table), sql.Identifier(areas_to_join), sql.Identifier(dataset))) #'-- and (st_area(st_intersection(b.geom, t.geom)) / st_area(b.geom)) > .10
+    if join_multiple_areas:
+        # run join
+        if join_data_is_areas:
+            cursor.execute(sql.SQL('SELECT b.%s, ARRAY_AGG(st_area(st_intersection(b.%s, t.geom)) / st_area(b.%s)) as coverage, ARRAY_AGG(t.geo_code) as lads, ARRAY_AGG(t.geo_code_gor) as gors INTO {} FROM ftables.{} t, {} b WHERE st_intersects(b.%s, t.%s) ORDER BY g.gid;').format(sql.Identifier(temp_table), sql.Identifier(areas_to_join), sql.Identifier(dataset)), [dataset_id_field, join_areas_geom_field, dataset_geom_field, dataset_geom_field, join_areas_geom_field])
+        else:
+            cursor.execute(sql.SQL('SELECT b.gid, ARRAY_AGG(t.geo_code) as lads, ARRAY_AGG(t.geo_code_gor) as gors INTO ___temp FROM ftables.{} t, {} b WHERE st_intersects(b.geom_, t.geom) GROUP BY b.gid;').format(sql.Identifier(areas_to_join), sql.Identifier(dataset)))
+
+        # create fields in dataset to join areas to
+        #for field in fields_to_join:
+        cursor.execute(sql.SQL('ALTER TABLE {} ADD IF NOT EXISTS lads character varying[];').format(sql.Identifier(dataset)))
+        cursor.execute(sql.SQL('ALTER TABLE {} ADD IF NOT EXISTS gors character varying[];').format(sql.Identifier(dataset)))
+
+        # add field for coverage values if join involves areas
+        if join_data_is_areas:
+            cursor.execute(sql.SQL('ALTER TABLE {} ADD IF NOT EXISTS lad_coverage float[];').format(sql.Identifier(dataset)))
+
+        # create index on the ids to speed up the update
+        cursor.execute(sql.SQL('CREATE INDEX %s ON {} USING btree(gid);').format(sql.Identifier(temp_table)), [temp_table+'_gid_idx'])
+        cursor.execute(sql.SQL('CREATE INDEX IF NOT EXISTS %s ON {} USING btree(gid);').format(sql.Identifier(dataset)), [dataset+'_gid_idx'])
+
+        # run the update
+        cursor.execute(sql.SQL('UPDATE {} as a SET lads = b.lads, gors = b.gors, lad_coverage = b.coverage FROM {} b WHERE a.gid = b.gid;').format(sql.Identifier(dataset), sql.Identifier(temp_table)))
+
+        # remote the temp tables
+        cursor.execute(sql.SQL('DROP INDEX %s;').format(sql.Identifier()), [temp_table+'_gid_idx'])
+        cursor.execute(sql.SQL('DROP TABLE {};').format(sql.Identifier(temp_table)))
+
+        # create index on new fields which are now populated
+        cursor.execute(sql.SQL('CREATE INDEX IF NOT EXISTS %s ON {} USING gin(lads);').format(sql.Identifier(dataset)), [dataset+'_lads_idx'])
+        cursor.execute(sql.SQL('CREATE INDEX IF NOT EXISTS %s ON {} USING gin(gors);').format(sql.Identifier(dataset)), [dataset+'_gor_idx'])
+
     else:
-        cursor.execute(sql.SQL('SELECT b.gid, ARRAY_AGG(t.geo_code) as lads, ARRAY_AGG(t.geo_code_gor) as gors INTO ___temp FROM ftables.{} t, {} b WHERE st_intersects(b.geom_, t.geom) GROUP BY b.gid;').format(sql.Identifier(areas_to_join), sql.Identifier(dataset)))
+        # if join result is to find a single area rather than multiple
 
-    # create fields in dataset to join areas to
-    #for field in fields_to_join:
-    cursor.execute(sql.SQL('ALTER TABLE {} ADD IF NOT EXISTS lads character varying[];').format(sql.Identifier(dataset)))
-    cursor.execute(sql.SQL('ALTER TABLE {} ADD IF NOT EXISTS gors character varying[];').format(sql.Identifier(dataset)))
+        # there should be some pre-processing in here before the spatial join runs
 
-    # add field for coverage values if join involves areas
-    if dataset_area:
-        cursor.execute(sql.SQL('ALTER TABLE {} ADD IF NOT EXISTS lad_coverage float[];').format(sql.Identifier(dataset)))
-
-    # create index on the ids to speed up the update
-    cursor.execute(sql.SQL('CREATE INDEX ___temp_gid_idx ON {} USING btree(gid);').format(sql.Identifier(temp_table)))
-    cursor.execute(sql.SQL('CREATE INDEX IF NOT EXISTS scheduledmonuments2019_gid_idx ON {} USING btree(gid);').format(sql.Identifier(dataset)))
-
-    # run the update
-    cursor.execute(sql.SQL('UPDATE {} as a SET lads = b.lads, gors = b.gors, lad_coverage = b.coverage FROM {} b WHERE a.gid = b.gid;').format(sql.Identifier(dataset), sql.Identifier(temp_table)))
-
-    # remote the temp tables
-    cursor.execute(sql.SQL('DROP INDEX ___temp_gid_idx;').format(sql.Identifier()))
-    cursor.execute(sql.SQL('DROP TABLE {};').format(sql.Identifier(temp_table)))
-
-    # create index on new fields which are now populated
-    cursor.execute(sql.SQL('CREATE INDEX IF NOT EXISTS %s ON {} USING gin(lads);').format(sql.Identifier(dataset)), [dataset+'_lads_idx'])
-    cursor.execute(sql.SQL('CREATE INDEX IF NOT EXISTS %s ON {} USING gin(gors);').format(sql.Identifier(dataset)), [dataset+'_gor_idx'])
+        # run spatial join
+        cursor.execute(sql.SQL('UPDATE {} a SET lad = b.lad_code, gor = b.gor_code FROM ftables.{} b WHERE ST_Contains(b.geom, a.centroid);').format(sql.Identifier(dataset), sql.Identifier(areas_to_join)))
